rename module to namespace in code
