+++ test/data/expectedNamespace.js
