+++ test/data/starWarsNamespace.js
